CTCF motifs on human chromosome 22 with example coverage.

This dataset is the same as motif.hg19.CTCF.chr22 but with one additional metadata colum, named "chip", holding ChIP-seq signals for all motifs in a windows of 1000 bp around the motif center as NumericList. The data is from a ChIP-seq experiment for STAT1 in human GM12878 cells. The full bigWig file can be downloaded from ENCODE (Dunham et al. 2012) http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeSydhTfbs/wgEncodeSydhTfbsGm12878Stat1StdSig.bigWig. See motif.hg19.CTCF and motif.hg19.CTCF.chr22 for a more details and the motif data set.

motif.hg19.CTCF.chr22.cov

Format

An object of class GRanges of length 917.

See also