Parse chromatin interactions from Tang et al. 2015 as GInteractions.

Reads pairwise ChIA-PET interaction from an input file.

parseLoopsTang(inFile, ...)

Arguments

inFile

input file with loops

...

additional arguments, that will be passed to GRanges functions.

Value

An GInteractions with loops from input file.

Details

It reads files with the following tab-delimited format:

chr12	48160351	48161634	chr12	48230665	48232848	27

chr7	77284664	77285815	chr7	77388242	77388928	7

chr4	128459961	128460166	chr4	128508304	128509082	4

This file format was used for ChIA-PET interaction data by Tang et al. 2015 http://dx.doi.org/10.1016/j.cell.2015.11.024. The last column of input file is added as annotation column with colname "score".

Examples

exampleLoopTang2015File <- system.file("extdata",
   "ChIA-PET_GM12878_Tang2015.chr22_1-30000000.clusters.txt",
   package = "sevenC")

gi <- parseLoopsTang(exampleLoopTang2015File)
#> INFO: Parse ChiA-PET interactions from file: /__w/_temp/Library/sevenC/extdata/ChIA-PET_GM12878_Tang2015.chr22_1-30000000.clusters.txt
#> Rows: 675 Columns: 7
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (2): X1, X4
#> dbl (5): X2, X3, X5, X6, X7
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

# read loops with custom seqinfo object:
customSeqInfo <- Seqinfo(seqnames = c("chr1", "chr22"),
   seqlengths = c(10^8, 10^8), isCircular = c(FALSE, FALSE),
   genome = "custom")
gi <- parseLoopsTang(exampleLoopTang2015File, seqinfo = customSeqInfo)
#> INFO: Parse ChiA-PET interactions from file: /__w/_temp/Library/sevenC/extdata/ChIA-PET_GM12878_Tang2015.chr22_1-30000000.clusters.txt
#> Rows: 675 Columns: 7
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (2): X1, X4
#> dbl (5): X2, X3, X5, X6, X7
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.